+++ src/main/java/net/silentchaos512/gear/data/part/PartsProvider.java
@@ -182,6 +182,13 @@ public class PartsProvider implements IDataProvider {
                     .stat(ItemStats.ATTACK_SPEED, -2.2f, StatInstance.Operation.ADD)
                     .stat(ItemStats.REPAIR_EFFICIENCY, 0.75f)
                     .stat(ItemStats.DURABILITY, -0.2f, StatInstance.Operation.MUL1);
+        if (isToolHead(builder, ModItems.KNIFE_BLADE))
+            return builder
+                    .stat(ItemStats.MELEE_DAMAGE, 1, StatInstance.Operation.ADD)
+                    .stat(ItemStats.ATTACK_SPEED, -1.6f, StatInstance.Operation.ADD)
+                    .stat(ItemStats.REPAIR_EFFICIENCY, 2)
+                    .stat(ItemStats.MELEE_DAMAGE, -0.5f, StatInstance.Operation.MUL1)
+                    .stat(ItemStats.DURABILITY, 0.25f, StatInstance.Operation.MUL1);
         if (isToolHead(builder, ModItems.MACHETE_BLADE))
             return builder
                     .stat(ItemStats.MELEE_DAMAGE, 4, StatInstance.Operation.ADD)